@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": ">=7.0.0",
+        "php": ">=7.1.0",
 
         "symfony/debug": "v3.4.18",
         "devfactory/minify": "1.0.7",
@@ -29,17 +29,18 @@
         "barryvdh/laravel-translation-manager": "v0.5.0",
         "chumper/zipper": "v1.0.2",
         "rachidlaasri/laravel-installer": "4.0.2",
-        "kitetail/zttp": "v0.4.0",
-        "rap2hpoutre/laravel-log-viewer": "v0.22.1",
+        "rap2hpoutre/laravel-log-viewer": "v2.0.0",
         "codedge/laravel-selfupdater": "1.4.3",
         "doctrine/dbal": "2.12.1",
         "egulias/email-validator": "2.1.10",
         "symfony/console": "v3.4.47",
-        "javoscript/laravel-macroable-models": "1.0.4"
+        "javoscript/laravel-macroable-models": "1.0.4",
+        "ramsey/uuid": "3.9.6",
+        "symfony/polyfill-mbstring": "v1.24.0"
     },
     "require-dev": {
         "barryvdh/laravel-debugbar": "v3.2.0",
-        "filp/whoops": "2.2.0",
+        "filp/whoops": "2.14.5",
         "fzaninotto/faker": "v1.9.2",
         "mockery/mockery": "1.1.0",
         "phpunit/phpunit": "6.5.8",
