@@ -12,7 +12,7 @@ return [
     | or any other location as required by the application or its packages.
     */
 
-    'version' => '1.7.27',
+    'version' => '1.8.8',
 
     /*
     |--------------------------------------------------------------------------
@@ -94,7 +94,6 @@ return [
     'locale'          => env('APP_LOCALE', 'en'),
     'locales'         => ['en', 'hr', 'da', 'nl', 'fr', 'de', 'ja', 'it', 'pl', 'pt-PT', 'pt-BR', 'ru', 'es', 'sk', 'sv','fa'],
     'default_locale'  => 'en',
-    'locales_rtl' => ['fa'],
 
     /*
     | app()->setLocale() in Localize middleware also changes config('app.locale'),
