@@ -21,13 +21,16 @@
         if (!isset($params)) {
             $params = [];
         }
+
+        // Sorting.
+        $sorting = App\Conversation::getConvTableSorting();
     @endphp
 
     @if (!request()->get('page'))
         @include('/conversations/partials/bulk_actions')
     @endif
 
-    <table class="table-conversations table @if (!empty($params['show_mailbox']))show-mailbox @endif" data-page="{{ (int)request()->get('page', 1) }}" @foreach ($params as $param_name => $param_value) data-param_{{ $param_name }}="{{ $param_value }}" @endforeach @if (!empty($conversations_filter)) @foreach ($conversations_filter as $filter_field => $filter_value) data-filter_{{ $filter_field }}="{{ $filter_value }}" @endforeach @endif >
+    <table class="table-conversations table @if (!empty($params['show_mailbox']))show-mailbox @endif" data-page="{{ (int)request()->get('page', 1) }}" @foreach ($params as $param_name => $param_value) data-param_{{ $param_name }}="{{ $param_value }}" @endforeach @if (!empty($conversations_filter)) @foreach ($conversations_filter as $filter_field => $filter_value) data-filter_{{ $filter_field }}="{{ $filter_value }}" @endforeach @endif @foreach ($sorting as $sorting_name => $sorting_value) data-param_{{ $sorting_name }}="{{ $sorting_value }}" @endforeach >
         <colgroup>
             {{-- todo: without this column table becomes not 100% wide --}}
             <col class="conv-current">
@@ -54,7 +57,11 @@
             @endif
             <th class="conv-attachment">&nbsp;</th>
             <th class="conv-subject" colspan="2">
-                <span>{{ __("Conversation") }}</span>
+                <span class="conv-col-sort" data-sort-by="subject" data-order="@if ($sorting['sort_by'] == 'subject'){{ $sorting['order'] }}@else{{ 'asc' }}@endif">
+                    {{ __("Conversation") }} 
+                     @if ($sorting['sort_by'] == 'subject' && $sorting['order'] =='desc')↑@endif
+                     @if ($sorting['sort_by'] == 'subject' && $sorting['order'] =='asc')↓@endif
+                </span>
             </th>
             @if ($show_assigned)
                 <th class="conv-owner">
@@ -71,18 +78,24 @@
             @endif
             @action('conversations_table.th_before_conv_number')
             <th class="conv-number">
-                <span>{{ __("Number") }}</span>
+                <span class="conv-col-sort" data-sort-by="number" data-order="@if ($sorting['sort_by'] == 'number'){{ $sorting['order'] }}@else{{ 'asc' }}@endif">
+                    {{ __("Number") }} 
+                     @if ($sorting['sort_by'] == 'number' && $sorting['order'] =='desc')↑@endif
+                     @if ($sorting['sort_by'] == 'number' && $sorting['order'] =='asc')↓@endif
+                </span>
             </th>
             <th class="conv-date">
                 <span>
-                    @if ($folder->type == App\Folder::TYPE_CLOSED){{ __("Closed") }}@elseif ($folder->type == App\Folder::TYPE_DRAFTS){{ __("Last Updated") }}@elseif ($folder->type == App\Folder::TYPE_DELETED){{ __("Deleted") }}@else{{ \Eventy::filter('conversations_table.column_title_date', __("Waiting Since"), $folder) }}@endif<i class="conv-sort">↓</i>
+                    <span class="conv-col-sort" data-sort-by="date" data-order="@if ($sorting['sort_by'] == 'date'){{ $sorting['order'] }}@else{{ 'asc' }}@endif">
+                        @if ($folder->type == App\Folder::TYPE_CLOSED){{ __("Closed") }}@elseif ($folder->type == App\Folder::TYPE_DRAFTS){{ __("Last Updated") }}@elseif ($folder->type == App\Folder::TYPE_DELETED){{ __("Deleted") }}@else{{ \Eventy::filter('conversations_table.column_title_date', __("Waiting Since"), $folder) }}@endif @if ($sorting['sort_by'] == 'date' && $sorting['order'] =='desc')↑@elseif ($sorting['sort_by'] == 'date' && $sorting['order'] =='asc')↓@elseif ($sorting['sort_by'] == '' && $sorting['order'] =='')↓@endif
+                    </a>
                 </span>
             </th>
           </tr>
         </thead>
         <tbody>
             @foreach ($conversations as $conversation)
-                <tr class="conv-row @if ($conversation->isActive()) conv-active @endif" data-conversation_id="{{ $conversation->id }}">
+                <tr class="conv-row @if ($conversation->isActive()) conv-active @endif @if ($conversation->isSpam()) conv-spam @endif" data-conversation_id="{{ $conversation->id }}">
                     @if (empty($no_checkboxes))<td class="conv-current">@if (!empty($viewers[$conversation->id]))
                                 <div class="viewer-badge @if (!empty($viewers[$conversation->id]['replying'])) viewer-replying @endif" data-toggle="tooltip" title="@if (!empty($viewers[$conversation->id]['replying'])){{ __(':user is replying', ['user' => $viewers[$conversation->id]['user']->getFullName()]) }}@else{{ __(':user is viewing', ['user' => $viewers[$conversation->id]['user']->getFullName()]) }}@endif"><div>
                             @endif</td>@else<td class="conv-current"></td>@endif
