@@ -4,5 +4,5 @@
 
 @section('message')
     {{ __('Access denied') }}
-    <br/><br/><small>{{ __('Go to') }} <a href="{{ route('dashboard') }}">{{ __('Homepage') }}</a></small>
+    <br/><br/><small>{{ __('Go to') }} <a href="{{ \Helper::urlHome() }}">{{ __('Homepage') }}</a></small>
 @endsection
