@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="{{ app()->getLocale() }}" @if(Helper::isLocaleRtl()) dir="rtl" @endif>
+<html lang="{{ app()->getLocale() }}">
 <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -8,7 +8,8 @@
     <!-- CSRF Token -->
     <meta name="csrf-token" content="{{ csrf_token() }}">
 
-    <title>@if ($__env->yieldContent('title_full'))@yield('title_full') @elseif ($__env->yieldContent('title'))@yield('title') - {{ config('app.name', 'FreeScout') }} @else{{ config('app.name', 'FreeScout') }}@endif</title>
+    <title>@if ($__env->yieldContent('title_full'))@yield('title_full') @elseif ($__env->yieldContent('title'))@yield('title')
+        - {{ config('app.name', 'FreeScout') }} @else{{ config('app.name', 'FreeScout') }}@endif</title>
 
     <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png">
     {{--<link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png">
@@ -24,12 +25,12 @@
     @php
         try {
             $styles= array('/css/fonts.css', '/css/bootstrap.css', '/css/select2/select2.min.css', '/js/featherlight/featherlight.min.css', '/js/featherlight/featherlight.gallery.min.css', '/css/magic-check.css', '/css/style.css' );
-            if (Helper::isLocaleRtl()) {
+            if (app()->getLocale() === "fa") {
                 $styles[] = '/css/bootstrap-rtl.css';
-                $styles[] = '/css/style-rtl.css';
+                $styles[] = '/css/rtl.css';
             }
     @endphp
-    {!! Minify::stylesheet(\Eventy::filter('stylesheets', $styles)) !!}
+    {!! Minify::stylesheet(\Eventy::filter('stylesheets',$styles)) !!}
     @php
         } catch (\Exception $e) {
             // Try...catch is needed to catch errors when activating a module and public symlink not created for module.
@@ -39,7 +40,9 @@
 
     @yield('stylesheets')
 </head>
-<body class="locale-{{ app()->getLocale() }} @if(Helper::isLocaleRtl()) rtl @endif @if (!Auth::user()) user-is-guest @endif @if (Auth::user() && Auth::user()->isAdmin()) user-is-admin @endif @yield('body_class')" @yield('body_attrs') @if (Auth::user()) data-auth_user_id="{{ Auth::user()->id }}" @endif>
+<body
+    class="@if (!Auth::user()) user-is-guest @endif @if (Auth::user() && Auth::user()->isAdmin()) user-is-admin @endif @yield('body_class')"
+    @yield('body_attrs') @if (Auth::user()) data-auth_user_id="{{ Auth::user()->id }}" @endif>
 <div id="app">
 
     @if (Auth::user() && empty(app('request')->x_embed) && empty($__env->yieldContent('guest_mode')))
@@ -49,7 +52,8 @@
                 <div class="navbar-header">
 
                     <!-- Collapsed Hamburger -->
-                    <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#app-navbar-collapse" aria-expanded="false">
+                    <button type="button" class="navbar-toggle collapsed" data-toggle="collapse"
+                            data-target="#app-navbar-collapse" aria-expanded="false">
                         <span class="sr-only">{{ __('Toggle Navigation') }}</span>
                         <span class="icon-bar"></span>
                         <span class="icon-bar"></span>
@@ -63,7 +67,7 @@
                         </a>
                     @else
                         <a class="navbar-brand" href="{{ route('dashboard') }}" title="{{ __('Dashboard') }}">
-                            <img src="@filter('layout.header_logo', asset('img/logo-brand.svg'))" height="100%" alt="" />
+                            <img src="@filter('layout.header_logo', asset('img/logo-brand.svg'))" height="100%" alt=""/>
                             {{-- config('app.name', 'FreeScout') --}}
                         </a>
                     @endif
@@ -77,15 +81,20 @@
                             $mailboxes = \Eventy::filter('menu.mailboxes', $mailboxes);
                         @endphp
                         @if (count($mailboxes) == 1)
-                            <li class="{{ \App\Misc\Helper::menuSelectedHtml('mailbox') }}"><a href="{{ \Eventy::filter('mailbox.url', route('mailboxes.view', ['id'=>$mailboxes[0]->id]), $mailboxes[0]) }}">{{ __('Mailbox') }}</a></li>
+                            <li class="{{ \App\Misc\Helper::menuSelectedHtml('mailbox') }}"><a
+                                    href="{{ \Eventy::filter('mailbox.url', route('mailboxes.view', ['id'=>$mailboxes[0]->id]), $mailboxes[0]) }}">{{ __('Mailbox') }}</a>
+                            </li>
                         @elseif (count($mailboxes) > 1)
                             <li class="dropdown {{ \App\Misc\Helper::menuSelectedHtml('mailbox') }}">
-                                <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-expanded="false" aria-haspopup="true" v-pre>
+                                <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button"
+                                   aria-expanded="false" aria-haspopup="true" v-pre>
                                     {{ __('Mailbox') }} <span class="caret"></span>
                                 </a>
                                 <ul class="dropdown-menu dm-scrollable">
                                     @foreach ($mailboxes as $mailbox_item)
-                                        <li @if ($mailbox_item->id == app('request')->id)class="active"@endif><a href="{{ \Eventy::filter('mailbox.url', route('mailboxes.view', ['id' => $mailbox_item->id]), $mailbox_item) }}">{{ $mailbox_item->name }}</a></li>
+                                        <li @if ($mailbox_item->id == app('request')->id)class="active"@endif><a
+                                                href="{{ \Eventy::filter('mailbox.url', route('mailboxes.view', ['id' => $mailbox_item->id]), $mailbox_item) }}">{{ $mailbox_item->name }}</a>
+                                        </li>
                                     @endforeach
                                 </ul>
                             </li>
@@ -96,26 +105,34 @@
                             || Eventy::filter('menu.manage.can_view', false)
                         )
                             <li class="dropdown {{ \App\Misc\Helper::menuSelectedHtml('manage') }}">
-                                <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-expanded="false" aria-haspopup="true" v-pre>
+                                <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button"
+                                   aria-expanded="false" aria-haspopup="true" v-pre>
                                     {{ __('Manage') }} <span class="caret"></span>
                                 </a>
 
                                 <ul class="dropdown-menu">
                                     @if (Auth::user()->isAdmin())
-                                        <li class="{{ \App\Misc\Helper::menuSelectedHtml('settings') }}"><a href="{{ route('settings') }}">{{ __('Settings') }}</a></li>
+                                        <li class="{{ \App\Misc\Helper::menuSelectedHtml('settings') }}"><a
+                                                href="{{ route('settings') }}">{{ __('Settings') }}</a></li>
                                     @endif
                                     @if (Auth::user()->can('viewMailboxMenu', Auth::user()))
-                                        <li class="{{ \App\Misc\Helper::menuSelectedHtml('mailboxes') }}"><a href="{{ route('mailboxes') }}">{{ __('Mailboxes') }}</a></li>
+                                        <li class="{{ \App\Misc\Helper::menuSelectedHtml('mailboxes') }}"><a
+                                                href="{{ route('mailboxes') }}">{{ __('Mailboxes') }}</a></li>
                                     @endif
                                     @action('menu.manage.after_mailboxes')
                                     @if (Auth::user()->isAdmin() || Auth::user()->hasPermission(App\User::PERM_EDIT_USERS))
-                                        <li class="{{ \App\Misc\Helper::menuSelectedHtml('users') }}"><a href="{{ route('users') }}">{{ __('Users') }}</a></li>
+                                        <li class="{{ \App\Misc\Helper::menuSelectedHtml('users') }}"><a
+                                                href="{{ route('users') }}">{{ __('Users') }}</a></li>
                                     @endif
                                     @if (Auth::user()->isAdmin())
-                                        <li class="{{ \App\Misc\Helper::menuSelectedHtml('modules') }}"><a href="{{ route('modules') }}">{{ __('Modules') }}</a></li>
-                                        <li class=""><a href="{{ asset('translations') }}">{{ __('Translate') }}</a></li>
-                                        <li class="{{ \App\Misc\Helper::menuSelectedHtml('logs') }}"><a href="{{ route('logs') }}">{{ __('Logs') }}</a></li>
-                                        <li class="{{ \App\Misc\Helper::menuSelectedHtml('system') }}"><a href="{{ route('system') }}">{{ __('System') }}</a></li>
+                                        <li class="{{ \App\Misc\Helper::menuSelectedHtml('modules') }}"><a
+                                                href="{{ route('modules') }}">{{ __('Modules') }}</a></li>
+                                        <li class=""><a href="{{ asset('translations') }}">{{ __('Translate') }}</a>
+                                        </li>
+                                        <li class="{{ \App\Misc\Helper::menuSelectedHtml('logs') }}"><a
+                                                href="{{ route('logs') }}">{{ __('Logs') }}</a></li>
+                                        <li class="{{ \App\Misc\Helper::menuSelectedHtml('system') }}"><a
+                                                href="{{ route('system') }}">{{ __('System') }}</a></li>
                                     @endif
                                     @action('menu.manage.append')
                                 </ul>
@@ -134,7 +151,10 @@
                                 @php
                                     $web_notifications_info = Auth::user()->getWebsiteNotificationsInfo();
                                 @endphp
-                                <a href="#" class="dropdown-toggle dropdown-toggle-icon @if ($web_notifications_info['unread_count']) @if ($web_notifications_info['unread_count']) has-unread @endif @endif" data-toggle="dropdown" role="button" aria-expanded="false" aria-haspopup="true" v-pre title="{{ __('Notifications') }}">
+                                <a href="#"
+                                   class="dropdown-toggle dropdown-toggle-icon @if ($web_notifications_info['unread_count']) @if ($web_notifications_info['unread_count']) has-unread @endif @endif"
+                                   data-toggle="dropdown" role="button" aria-expanded="false" aria-haspopup="true" v-pre
+                                   title="{{ __('Notifications') }}">
                                     <i class="glyphicon glyphicon-bell"></i>
                                 </a>
 
@@ -143,9 +163,14 @@
                                         <div class="web-notifications-header">
                                             <h1>
                                                 {{ __('Notifications') }}
-                                                <small class="web-notifications-count  @if (!(int)$web_notifications_info['unread_count']) hidden @endif" title="{{ __('Unread Notifications') }}" data-toggle="tooltip">@if ($web_notifications_info['unread_count']){{ $web_notifications_info['unread_count'] }}@endif</small>
+                                                <small
+                                                    class="web-notifications-count  @if (!(int)$web_notifications_info['unread_count']) hidden @endif"
+                                                    title="{{ __('Unread Notifications') }}"
+                                                    data-toggle="tooltip">@if ($web_notifications_info['unread_count']){{ $web_notifications_info['unread_count'] }}@endif</small>
                                             </h1>
-                                            <a href="#" class="web-notifications-mark-read @if (!(int)$web_notifications_info['unread_count']) hidden @endif" data-loading-text="{{ __('Processing') }}…">
+                                            <a href="#"
+                                               class="web-notifications-mark-read @if (!(int)$web_notifications_info['unread_count']) hidden @endif"
+                                               data-loading-text="{{ __('Processing') }}…">
                                                 {{ __('Mark all as read') }}
                                             </a>
                                         </div>
@@ -159,7 +184,8 @@
 
                                                 @if ($web_notifications_info['notifications']->hasMorePages())
                                                     <li class="web-notification-more">
-                                                        <button class="btn btn-link btn-block link-dark" data-loading-text="{{ __('Loading') }}…">
+                                                        <button class="btn btn-link btn-block link-dark"
+                                                                data-loading-text="{{ __('Loading') }}…">
                                                             {{ __('Load more') }}
                                                         </button>
                                                     </li>
@@ -182,12 +208,19 @@
 
                             <li class="dropdown">
 
-                                <a href="#" class="dropdown-toggle dropdown-toggle-icon dropdown-toggle-account" data-toggle="dropdown" role="button" aria-expanded="false" aria-haspopup="true" v-pre title="{{ __('Account') }}" aria-label="{{ __('Account') }}">
-                                    <span class="photo-sm">@include('partials/person_photo', ['person' => Auth::user()])</span>&nbsp;<span class="nav-user">{{ Auth::user()->first_name }}</span> <span class="caret"></span>
+                                <a href="#" class="dropdown-toggle dropdown-toggle-icon dropdown-toggle-account"
+                                   data-toggle="dropdown" role="button" aria-expanded="false" aria-haspopup="true" v-pre
+                                   title="{{ __('Account') }}" aria-label="{{ __('Account') }}">
+                                    <span
+                                        class="photo-sm">@include('partials/person_photo', ['person' => Auth::user()])</span>&nbsp;<span
+                                        class="nav-user">{{ Auth::user()->first_name }}</span> <span
+                                        class="caret"></span>
                                 </a>
 
                                 <ul class="dropdown-menu">
-                                    <li><a href="{{ route('users.profile', ['id'=>Auth::user()->id]) }}">{{ __('Your Profile') }}</a></li>
+                                    <li>
+                                        <a href="{{ route('users.profile', ['id'=>Auth::user()->id]) }}">{{ __('Your Profile') }}</a>
+                                    </li>
                                     @action('menu_right.user.after_profile')
                                     <li class="divider"></li>
                                     <li>
@@ -197,37 +230,52 @@
                                             {{ __('Log Out') }}
                                         </a>
 
-                                        <form id="logout-form" action="{{ route('logout') }}" method="POST" style="display: none;">
+                                        <form id="logout-form" action="{{ route('logout') }}" method="POST"
+                                              style="display: none;">
                                             {{ csrf_field() }}
                                         </form>
                                     </li>
                                     <li class="divider hidden in-app-switcher"></li>
                                     <li>
-                                        <a href="javascript:switchHelpdeskUrl();void(0);" class="hidden in-app-switcher">{{ __('Switch Helpdesk URL') }}</a>
+                                        <a href="javascript:switchHelpdeskUrl();void(0);"
+                                           class="hidden in-app-switcher">{{ __('Switch Helpdesk URL') }}</a>
                                     </li>
                                 </ul>
                             </li>
 
                             <li class="dropdown">
-                                <a href="#" class="dropdown-toggle dropdown-toggle-icon" data-toggle="dropdown" role="button" aria-expanded="false" aria-haspopup="true" v-pre title="{{ __('Search') }}" aria-label="{{ __('Search') }}" id="search-dt">
+                                <a href="#" class="dropdown-toggle dropdown-toggle-icon" data-toggle="dropdown"
+                                   role="button" aria-expanded="false" aria-haspopup="true" v-pre
+                                   title="{{ __('Search') }}" aria-label="{{ __('Search') }}" id="search-dt">
                                     <i class="glyphicon glyphicon-search"></i>
                                 </a>
 
                                 <ul class="dropdown-menu dropdown-with-icons">
                                     <li>
-                                        <form class="form-inline form-nav-search" role="form" action="{{ route('conversations.search') }}" target="_blank">
+                                        <form class="form-inline form-nav-search" role="form"
+                                              action="{{ route('conversations.search') }}" target="_blank">
                                             <div class="input-group">
                                                 <input type="text" class="form-control" name="q">
                                                 <span class="input-group-btn">
-                                                        <button class="btn btn-default" type="submit">{{ __('Search') }}</button>
+                                                        <button class="btn btn-default"
+                                                                type="submit">{{ __('Search') }}</button>
                                                     </span>
                                             </div>
                                         </form>
                                     </li>
-                                    <li><a href="{{ route('conversations.search', ['f' => ['following' => 'yes']]) }}"><i class="glyphicon glyphicon-chevron-right"></i> {{ __("Conversations I'm following") }}</a></li>
-                                    <li><a href="{{ route('conversations.search', ['f' => ['assigned' => Auth::user()->id, 'status' => [App\Conversation::STATUS_ACTIVE, App\Conversation::STATUS_PENDING]]]) }}"><i class="glyphicon glyphicon-chevron-right"></i> {{ __('My open conversations') }}</a></li>
+                                    <li>
+                                        <a href="{{ route('conversations.search', ['f' => ['following' => 'yes']]) }}"><i
+                                                class="glyphicon glyphicon-chevron-right"></i> {{ __("Conversations I'm following") }}
+                                        </a></li>
+                                    <li>
+                                        <a href="{{ route('conversations.search', ['f' => ['assigned' => Auth::user()->id, 'status' => [App\Conversation::STATUS_ACTIVE, App\Conversation::STATUS_PENDING]]]) }}"><i
+                                                class="glyphicon glyphicon-chevron-right"></i> {{ __('My open conversations') }}
+                                        </a></li>
                                     @if (in_array(Route::currentRouteName(), ['mailboxes.view', 'mailboxes.view.folder']))
-                                        <li><a href="{{ route('conversations.search', ['f' => ['mailbox' => app('request')->id]]) }}"><i class="glyphicon glyphicon-chevron-right"></i> {{ __("All from current mailbox") }}</a></li>
+                                        <li>
+                                            <a href="{{ route('conversations.search', ['f' => ['mailbox' => app('request')->id]]) }}"><i
+                                                    class="glyphicon glyphicon-chevron-right"></i> {{ __("All from current mailbox") }}
+                                            </a></li>
                                     @endif
                                 </ul>
                             </li>
@@ -257,12 +305,15 @@
         && empty(app('request')->x_embed) && empty($__env->yieldContent('no_footer')))
         <div class="footer">
             @if (!\Eventy::filter('footer.text', ''))
-            &copy; 2018-{{ date('Y') }} <a href="{{ config('app.freescout_url') }}" target="blank">{{ \Config::get('app.name') }}</a> — {{ __('Free open source help desk & shared mailbox') }}
+            &copy; 2018-{{ date('Y') }} <a href="{{ config('app.freescout_url') }}"
+                                           target="blank">{{ \Config::get('app.name') }}</a>
+            — {{ __('Free open source help desk & shared mailbox') }}
             @else
                 {!! \Eventy::filter('footer.text', '') !!}
             @endif
             @if (!Auth::user())
-                <a href="javascript:switchHelpdeskUrl();void(0);" class="hidden in-app-switcher"><br/>{{ __('Switch Helpdesk URL') }}</a>
+                <a href="javascript:switchHelpdeskUrl();void(0);"
+                   class="hidden in-app-switcher"><br/>{{ __('Switch Helpdesk URL') }}</a>
             @endif
             {{-- Show version to admin only --}}
             @if (Auth::user() && Auth::user()->isAdmin())
